refactor(gateway): move event wire types to core
